Fix problem with getOperatorName, with response of AT+COPS?
The +COPS response has optional fields and we didn't account for that.

Also split up the work of doing SIm check, enabling auto registration.
Connect is now also waiting for a minimum RSSI. Currently hardcoded as -93.
Notice that there is a TODO in enableAutoRegistration to remove the delay
and the readResponse.