Improved interface for MQTT
There are a few extra functions that are needed for the upcoming
changes in Sodaq_MQTT. But there is an API change of receiveMQTTPacket.

There are several bug fixes and error checking improvements.

To mention one bug fix in particular is that URCs do not cause a
return in readResponse anymore.

A new feature was added to temporarily disable diag output while doing
isAlive. This reduces unnecessary diag while polling for incoming
packets.

Another added feature is a call back handler for when a socket is
closed by the remote (and causes a +UUSOCL).