Rename setMinSignalQuality to setMinRSSI
This is to avoid confusion what "signal quality" actually means. There
are two other terms, RSSI and CSQ, and we will now use RSSI if we can.
For convenience there is a setMinCSQ function.

Notice that there are two conversion functions, convertCSQ2RSSI and
convertRSSI2CSQ

The 3Gbee component was not yet using the _minRSSI setting. That has
been corrected as well.